Renamed choosed resux branch to propperty.
