+++ src/hocs/with-feedback.e2e.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import {configure, shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import {withFeedback} from "./with-feedback.jsx";
-import {commentsSelector, formStatusSelector} from '../redux/choosed/choosed-selector';
+import {commentsSelector, formStatusSelector} from '../redux/property/property-selectors.js';
 
 configure({adapter: new Adapter()});
 
@@ -10,7 +10,7 @@ jest.mock(`react-redux`, () => ({
   useSelector: jest.fn((fn) => fn()),
   useDispatch: jest.fn(),
 }));
-jest.mock(`../redux/choosed/choosed-selector`);
+jest.mock(`../redux/property/property-selectors`);
 
 
 const TestComponent = () => <div>Somw text.</div>;
