--- src/redux/choosed/choosed-selector.js
@@ -1,5 +1,0 @@
-
-export const commentsSelector = (state) => state.choosed.comments;
-export const formStatusSelector = (state) => state.choosed.isFormSending;
-export const nearbySelector = (state) => state.choosed.nearby;
-export const choosedSelector = (state) => state.choosed.choosed;
