+++ src/redux/property/property-reducer.test.js
@@ -1,9 +1,9 @@
-import {choosedReducer} from './choosed-reducer.js';
+import {propertyReducer} from './property-reducer.js';
 import {SET_COMMENTS} from "./types.js";
 
 describe(`choosedReducer`, () => {
   it(`should SET_COMMENTS.`, () => {
-    const state = choosedReducer(
+    const state = propertyReducer(
         {},
         {
           type: SET_COMMENTS,
