+++ src/redux/property/property-selectors.js
@@ -1,0 +1,5 @@
+
+export const commentsSelector = (state) => state.property.comments;
+export const formStatusSelector = (state) => state.property.isFormSending;
+export const nearbySelector = (state) => state.property.nearby;
+export const choosedSelector = (state) => state.property.choosed;
