+++ src/redux/property/types.js
